+++ teku/src/main/java/tech/pegasys/teku/cli/options/ValidatorOptions.java
@@ -146,6 +146,20 @@ public class ValidatorOptions {
   private boolean blockV3Enabled = ValidatorConfig.DEFAULT_BLOCK_V3_ENABLED;
 
   @Option(
+      names = {"--Xvalidate-locally-created-blocks"},
+      paramLabel = "<BOOLEAN>",
+      description =
+          """
+             Enable gossip validation when publishing a block which has been created locally.
+             This is only applicable when "gossip" broadcast validation level has been requested,""",
+      hidden = true,
+      showDefaultValue = CommandLine.Help.Visibility.ALWAYS,
+      arity = "0..1",
+      fallbackValue = "true")
+  private boolean validateLocallyCreatedBlocks =
+      ValidatorConfig.DEFAULT_VALIDATE_LOCALLY_CREATED_BLOCKS;
+
+  @Option(
       names = {"--exit-when-no-validator-keys-enabled"},
       paramLabel = "<BOOLEAN>",
       description = "Enable terminating the process if no validator keys are found during startup",
@@ -194,6 +208,7 @@ public class ValidatorOptions {
                 .doppelgangerDetectionEnabled(doppelgangerDetectionEnabled)
                 .executorThreads(executorThreads)
                 .blockV3enabled(blockV3Enabled)
+                .validateLocallyCreatedBlocks(validateLocallyCreatedBlocks)
                 .exitWhenNoValidatorKeysEnabled(exitWhenNoValidatorKeysEnabled)
                 .shutdownWhenValidatorSlashedEnabled(shutdownWhenValidatorSlashed));
     validatorProposerOptions.configure(builder);