+++ validator/api/src/main/java/tech/pegasys/teku/validator/api/ValidatorConfig.java
@@ -46,6 +46,7 @@ public class ValidatorConfig {
   public static final boolean DEFAULT_FAILOVERS_SEND_SUBNET_SUBSCRIPTIONS_ENABLED = true;
   public static final boolean DEFAULT_FAILOVERS_PUBLISH_SIGNED_DUTIES_ENABLED = true;
   public static final boolean DEFAULT_BLOCK_V3_ENABLED = false;
+  public static final boolean DEFAULT_VALIDATE_LOCALLY_CREATED_BLOCKS = false;
   public static final boolean DEFAULT_EXIT_WHEN_NO_VALIDATOR_KEYS_ENABLED = false;
   public static final boolean DEFAULT_VALIDATOR_CLIENT_SSZ_BLOCKS_ENABLED = true;
   public static final boolean DEFAULT_VALIDATOR_CLIENT_USE_POST_VALIDATORS_ENDPOINT_ENABLED = true;
@@ -97,6 +98,7 @@ public class ValidatorConfig {
   private final boolean failoversSendSubnetSubscriptionsEnabled;
   private final boolean failoversPublishSignedDutiesEnabled;
   private final boolean blockV3Enabled;
+  private final boolean validateLocallyCreatedBlocks;
   private final boolean exitWhenNoValidatorKeysEnabled;
   private final boolean shutdownWhenValidatorSlashedEnabled;
   private final UInt64 builderRegistrationDefaultGasLimit;
@@ -140,6 +142,7 @@ public class ValidatorConfig {
       final boolean failoversSendSubnetSubscriptionsEnabled,
       final boolean failoversPublishSignedDutiesEnabled,
       final boolean blockV3Enabled,
+      final boolean validateLocallyCreatedBlocks,
       final boolean exitWhenNoValidatorKeysEnabled,
       final boolean shutdownWhenValidatorSlashedEnabled,
       final UInt64 builderRegistrationDefaultGasLimit,
@@ -183,6 +186,7 @@ public class ValidatorConfig {
     this.failoversSendSubnetSubscriptionsEnabled = failoversSendSubnetSubscriptionsEnabled;
     this.failoversPublishSignedDutiesEnabled = failoversPublishSignedDutiesEnabled;
     this.blockV3Enabled = blockV3Enabled;
+    this.validateLocallyCreatedBlocks = validateLocallyCreatedBlocks;
     this.exitWhenNoValidatorKeysEnabled = exitWhenNoValidatorKeysEnabled;
     this.shutdownWhenValidatorSlashedEnabled = shutdownWhenValidatorSlashedEnabled;
     this.builderRegistrationDefaultGasLimit = builderRegistrationDefaultGasLimit;
@@ -324,6 +328,10 @@ public class ValidatorConfig {
     return blockV3Enabled;
   }
 
+  public boolean isValidateLocallyCreatedBlocksEnabled() {
+    return validateLocallyCreatedBlocks;
+  }
+
   public boolean isExitWhenNoValidatorKeysEnabled() {
     return exitWhenNoValidatorKeysEnabled;
   }
@@ -404,6 +412,7 @@ public class ValidatorConfig {
     private boolean failoversPublishSignedDutiesEnabled =
         DEFAULT_FAILOVERS_PUBLISH_SIGNED_DUTIES_ENABLED;
     private boolean blockV3Enabled = DEFAULT_BLOCK_V3_ENABLED;
+    private boolean validateLocallyCreatedBlocks = DEFAULT_VALIDATE_LOCALLY_CREATED_BLOCKS;
     private boolean exitWhenNoValidatorKeysEnabled = DEFAULT_EXIT_WHEN_NO_VALIDATOR_KEYS_ENABLED;
     private boolean shutdownWhenValidatorSlashedEnabled =
         DEFAULT_SHUTDOWN_WHEN_VALIDATOR_SLASHED_ENABLED;
@@ -609,6 +618,11 @@ public class ValidatorConfig {
       return this;
     }
 
+    public Builder validateLocallyCreatedBlocks(final boolean validateLocallyCreatedBlocks) {
+      this.validateLocallyCreatedBlocks = validateLocallyCreatedBlocks;
+      return this;
+    }
+
     public Builder exitWhenNoValidatorKeysEnabled(final boolean exitWhenNoValidatorKeysEnabled) {
       this.exitWhenNoValidatorKeysEnabled = exitWhenNoValidatorKeysEnabled;
       return this;
@@ -704,6 +718,7 @@ public class ValidatorConfig {
           failoversSendSubnetSubscriptionsEnabled,
           failoversPublishSignedDutiesEnabled,
           blockV3Enabled,
+          validateLocallyCreatedBlocks,
           exitWhenNoValidatorKeysEnabled,
           shutdownWhenValidatorSlashedEnabled,
           builderRegistrationDefaultGasLimit,